refactor(guard): keep connect first-sync helper in a small module
Move shared first-sync failure recording out of the already oversized connect flow module so the quality ratchet stays green.
